[FIX] fix test error, by separate test view into another class
